Feature/add features skip display and xlsx password (#2)
* first commit

* Update README.md

* Update README.md

* Merge branch 'master' into develop

* initial update

* initial update: modal and encryption generator

* updated modal show with global event toggle show in store

* initial update for input file

* added calculation and decrypt file and download result

* added catch error for downloading result and added new worksheet for result answer and done

* added default env

* Update README.md

* updated no skip answer and clean package json

* enhanced button from landing page
