+++ frontend/README.md
@@ -1,46 +1,36 @@
-# Getting Started with Create React App
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+# BPI PsychoTest Questionaire App
 
-## Available Scripts
+A comprehensive digital solution for administering psychological tests. This project integrates a flexible and user-friendly questionnaire system into a paperless environment, making it easier to manage, administer, and analyze psychological assessments without the need for physical paperwork.
 
-In the project directory, you can run:
+## Features
 
-### `npm start`
+- **Paperless Integration**: Designed to work within a paperless system for easy distribution and data collection.
 
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
+## Installation
 
-The page will reload if you make edits.\
-You will also see any lint errors in the console.
+### Requirements
 
-### `npm test`
+[Tauri](https://tauri.app/)
 
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
+```
+// Running frontend local
 
-### `npm run build`
+git clone https://github.com/debuggingeveryday/bpi-questionaire.git
+cd frontend
+npm install
 
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
+// When exporting to software application
+/* From main direcory
+npm intall
+npm run tauri dev
+/* For build
+npm run tauri build
+```
 
-The build is minified and the filenames include the hashes.\
-Your app is ready to be deployed!
+# Testing
 
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+```
 
-### `npm run eject`
+```
 
-**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
-
-If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
-
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
-
-You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
-
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
