+++ frontend/package.json
@@ -8,9 +8,14 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/crypto-js": "^4.2.2",
+    "@types/file-saver": "^2.0.7",
     "@types/jest": "^27.5.2",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
+    "crypto-js": "^4.2.0",
+    "dotenv": "^16.4.5",
+    "file-saver": "^2.0.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
@@ -50,6 +55,7 @@
     "@types/react-transition-group": "^4.4.10",
     "autoprefixer": "^10.4.19",
     "cypress": "^13.13.3",
+    "cypress-file-upload": "^5.0.8",
     "postcss": "^8.4.40",
     "react-icons": "^5.2.1",
     "react-router-dom": "^6.26.0",
