+++ frontend/src/Components/Modal.tsx
@@ -1,0 +1,25 @@
+import * as React from 'react';
+
+interface IProps {
+  show: boolean;
+  children: React.ReactNode;
+}
+
+const Modal: React.FunctionComponent<IProps> = ({ show, children }): JSX.Element => {
+  return (
+    <div className={`relative z-10 ${!show ? 'hidden' : ''}`}>
+      <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity"></div>
+      <div className="fixed inset-0 z-10 w-screen overflow-y-auto">
+        <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
+          <div className="relative transform overflow-hidden rounded-lg bg-white text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg">
+            <div className="bg-white px-4 pb-4 pt-5 sm:p-6 sm:pb-4">
+              {children}
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Modal;
