--- frontend/src/Pages/Finished/Finished.tsx
@@ -1,12 +1,0 @@
-
-export default function Finished () {
-
-  return (
-    <div className="h-screen grid mx-auto justify-center">
-      <div className="w-1/2 place-self-center">
-        <h1 className="text-center text-5xl">Done</h1>
-        <button className="text-2xl bg-gray-900 text-white p-4 px-8 rounded-lg">Submit</button>
-      </div>
-    </div>
-  )
-}
