+++ frontend/src/Pages/Questionaire/Questionaire.tsx
@@ -2,9 +2,10 @@ import React, { useState, useEffect } from 'react';
 import { useParams, Link, useNavigate } from 'react-router-dom';
 import { IoArrowForwardSharp } from "react-icons/io5";
 import { IoMdArrowBack } from "react-icons/io";
-import * as XLSX from 'xlsx'
 import { useStoreContext } from '../../Store/Store'
 import { FaHome } from "react-icons/fa";
+import { saveAs } from 'file-saver'
+import CryptoJS from 'crypto-js';
 
 // TODO: create constants directory and file
 const STYLE_CLASS = {
@@ -23,7 +24,6 @@ function Questionaire() {
   const [showPrevious, setShowPrevious] = useState(false)
   const [showSubmit, setShowSubmit] = useState(false)
   const {
-    test,
     question,
     updateShowAlert,
     statusQuestion,
@@ -85,14 +85,12 @@ function Questionaire() {
 
   useEffect(() => {
     initializeQuestionaire()
-
-    console.log("on load Questionaire.tsx")
     return () => console.log("clean up")
   }, [])
 
 
   const next = (() => {
-    if (id) navigate(`/questionaire/${parseInt(id) + 1}`)
+    if (id && isDirty) navigate(`/questionaire/${parseInt(id) + 1}`)
   })
 
   const back = (() => {
@@ -109,11 +107,7 @@ function Questionaire() {
         answer
       }
     ))
-
-    const worksheet = XLSX.utils.json_to_sheet(rows);
-    const workbook = XLSX.utils.book_new();
-    XLSX.utils.book_append_sheet(workbook, worksheet, "Exported_Result");
-    
+  
     // TODO: make a util and put this to time.ts
     const date = new Date();
     const [month, day, year] = [
@@ -128,10 +122,12 @@ function Questionaire() {
       date.getSeconds(),
     ];
 
-    const fileName = `${year}${month}${day}${hour}${minutes}${seconds}_bpi-result.xlsx`
-
-    XLSX.writeFile(workbook, fileName, { compression: true });
-
+    const passphrase: any = process.env.REACT_APP_PASSPHASE;
+    const encryptedData = await CryptoJS.AES.encrypt(JSON.stringify(rows), passphrase).toString();
+    const blob = new Blob([encryptedData], {type: "text/plain;charset=utf-8"});
+    const fileName = `${month}${day}${year}${hour}${minutes}${seconds}_bpi-result.txt`
+    saveAs(blob, fileName);
+   
     updateShowAlert({
       show: true,
       color: 'bg-green-400',
@@ -144,38 +140,35 @@ function Questionaire() {
 
   return (
     <>
-    <Link className="ml-5 hover:underline flex" to={"/"}><FaHome className="mt-1 mr-1" />Home</Link>
-    <h1><pre>{JSON.stringify(test, null, 2)}</pre></h1>
-    {/* 
-    <div className="absolute"><pre>{JSON.stringify(question, null, 2)}</pre></div>
-    <button type="button" onClick={() => updateTest(1)}>update</button>
-    */}
-    <div className="h-screen grid">
-      <div className="w-2/3 mx-auto place-self-center">
-        <div className="grid grid-cols-2 justify-between">
-          <div className="place-self-start">
-            <button className={`flex text-2xl hover:underline ${showPrevious ? "hidden" : ""}`} onClick={back}><IoMdArrowBack className="mt-1" />Previous</button>
+      <Link className="ml-5 hover:underline flex" to={"/"}><FaHome className="mt-1 mr-1" />Home</Link>
+      <div className="h-screen grid">
+        <div className="w-2/3 mx-auto place-self-center">
+          <div className="grid grid-cols-2 justify-between">
+            <div className="place-self-start">
+              <button className={`flex text-2xl hover:underline ${showPrevious ? "hidden" : ""}`} onClick={back}><IoMdArrowBack className="mt-1" />Previous</button>
+            </div>
+            <div className="place-self-end">
+              {isDirty && (
+                <button className={`flex text-2xl hover:underline ${showNext ? "hidden" : ""}`} onClick={next}>Next<IoArrowForwardSharp className="mt-1" /></button>
+              )}
+            </div>
           </div>
-          <div className="place-self-end">
-            <button className={`flex text-2xl hover:underline ${showNext ? "hidden" : ""}`} onClick={next}>Next<IoArrowForwardSharp className="mt-1" /></button>
-          </div>
-        </div>
 
-        <div className="grid grid-cols-2 min-w-1/8 p-8 gap-y-16">
-          <h1 className="text-5xl col-span-2 font-bold text-center text-gray-500">{questionTitle}</h1>
-        </div>
-        <div className="grid grid-cols-2 gap-x-16">
-          <button className={`px-8 py-4 m-1 hover:m-0 text-2xl justify-self-end text-gray-900 focus:outline-none rounded-full border border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-4 focus:ring-gray-100 ${isDirty && answer ? STYLE_CLASS.activeButton : STYLE_CLASS.nonActiveButton }`} onClick={() => answerQuestionaire(true)}><span>True</span></button>
-          <button className={`px-8 py-4 m-1 hover:m-0 text-2xl justify-self-start text-gray-900 focus:outline-none rounded-full border border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-4 focus:ring-gray-100 ${isDirty && !answer ? STYLE_CLASS.activeButton : STYLE_CLASS.nonActiveButton }`} onClick={() => answerQuestionaire(false)}><span>False</span></button>
-        </div>
-        <div className="flex space-x-2">
-          {statusQuestion && (
-            <h1 className="text-2xl">{statusQuestion.countAnswered}/{statusQuestion.total}</h1>
-          )}
-          <button type="button" onClick={exportFile} className={`text-2xl hover:underline ${!showSubmit ? "hidden" : ""}`}>Export</button>
+          <div className="grid grid-cols-2 min-w-1/8 p-8 gap-y-16">
+            <h1 className="text-5xl col-span-2 font-bold text-center text-gray-500">{questionTitle}</h1>
+          </div>
+          <div className="grid grid-cols-2 gap-x-16">
+            <button className={`px-8 py-4 m-1 hover:m-0 text-2xl justify-self-end text-gray-900 focus:outline-none rounded-full border border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-4 focus:ring-gray-100 ${isDirty && answer ? STYLE_CLASS.activeButton : STYLE_CLASS.nonActiveButton }`} onClick={() => answerQuestionaire(true)}><span>True</span></button>
+            <button className={`px-8 py-4 m-1 hover:m-0 text-2xl justify-self-start text-gray-900 focus:outline-none rounded-full border border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-4 focus:ring-gray-100 ${isDirty && !answer ? STYLE_CLASS.activeButton : STYLE_CLASS.nonActiveButton }`} onClick={() => answerQuestionaire(false)}><span>False</span></button>
+          </div>
+          <div className="flex space-x-2">
+            {statusQuestion && (
+              <h1 className="text-2xl">{statusQuestion.countAnswered}/{statusQuestion.total}</h1>
+            )}
+            <button type="button" onClick={exportFile} className={`text-2xl hover:underline ${!showSubmit ? "hidden" : ""}`}>Export</button>
+          </div>
         </div>
       </div>
-    </div>
     </>
   );
 }
