+++ frontend/src/Router.tsx
@@ -9,12 +9,10 @@ import { CSSTransition, SwitchTransition } from 'react-transition-group'
 import Welcome from './Pages/Welcome/Welcome'
 import Questionaire from './Pages/Questionaire/Questionaire'
 import NotFound from './NotFound'
-import Finished from './Pages/Finished/Finished'
 
 const routes = [
   { path: '/', name: 'Welcome', element: <Welcome />, nodeRef: createRef() },
   { path: '/questionaire/:id', name: 'Questionaire', element: <Questionaire />, nodeRef: createRef() },
-  { path: '/finished', name: 'Finished', element: <Finished />, nodeRef: createRef() },
   { path: '*', name: 'Not Found', element: <NotFound />, nodeRef: createRef() }
 ]
 
@@ -25,9 +23,6 @@ const Router = () => {
 
   return (
     <>
-      <div className="grid grid-cols-2">
-        {/* <button type="button" className="justify-self-start text-2xl flex" onClick={() => navigate(-1)}><IoMdArrowBack className="mt-1 text-2xl" />Back</button> */}
-      </div>
       <SwitchTransition>
         <CSSTransition
           key={location.pathname}
