+++ frontend/src/constants/questionaire.ts
@@ -1,0 +1,244 @@
+const QUESTIONAIRES: Array<string> = [
+'It\'s easy for me to keep fit and healthy. ',
+'My present situation is hopeless. ',
+'I care about what other people think of me. ',
+'Sometimes I feel like smashing things. ',
+'I would enjoy betting on horses. ',
+'No one is making things work out badly for me. ',
+'I feel frightened when I have to go out alone. ',
+'I never have visions in which imaginary people or things appear. ',
+'Many times I act without thinking. ',
+'I enjoy being with people. ',
+'I have given up hope of ever amounting to anything. ',
+'I frequently think of the same silly thing over and over for hours. ',
+'Sometimes my legs lose their strength so that I can\'t walk. ',
+'I rarely feel disappointed. ',
+'Very few things excite me. ',
+'I would never intentionally hurt someone\'s feelings. ',
+'For the most part people are honest. ',
+'Someone has stolen my free will. ',
+'Even at the end of a hard day, I remain relaxed and at ease. ',
+'I seem to hear an unknown voice wherever I go. ',
+'I would not do something foolhardy just for the fun of it. ',
+'I keep my distance from other people. ',
+'I deserve my share of good luck. ',
+'There have been days when I have done things without being able to recall anything at all about them. ',
+'I am free of aches and pains. ',
+'There is not much to be interested in anymore. ',
+'Some movies make me quite happy or sad. ',
+'No one gets away with insulting me. ',
+'I have been in trouble with the law more than once. ',
+'I rarely feel that someone is trying to get the best of me. ',
+'Although I really try, I cannot stop feeling tense. ',
+'I do not experience peculiar voices warning me of danger. ',
+'I often behave in a reckless manner. ',
+'I like to speak to strangers. ',
+'I am only suited for the lowest and most simple sort of work. ',
+'I sometimes have fits and seizures that I cannot control. ',
+'My stomach is easily upset. ',
+'My future is cheery. ',
+'I would not be tempted by a promise of getting something for nothing. ',
+'I can get along quite well with irritable people. ',
+'No matter how easy or safe it was, I would never steal money. ',
+'I can tell that someone has searched through my possessions a number of times. ',
+'I remain quite cool when things go badly. ',
+'Fancy colored lights sometimes float through my brain. ',
+'Ideas do not race through my head faster than I can speak them. ',
+'Most of the time I prefer to be alone. ',
+'I am the type of person who can be relied upon. ',
+'I don\'t think my life is worth living. ',
+'I seldom have any bodily discomfort. ',
+'I live a gloomy and boring life. ',
+'At times I say things about my friends that aren\'t nice. ',
+'If someone does something I dislike, I usually tell that person about it. ',
+'I would enjoy cheating certain people. ',
+'I never have the feeling that someone is out to do away with me. ',
+'It frightens me to think about things that bother me. ',
+'Ordinary things never appear "foggy" or far away to me. ',
+'Sometimes I suddenly get up and act without warning or reason. ',
+'I would rather work with a group of people than by myself. ',
+'I am no good to anyone. ',
+'I have nightmares almost every night. ',
+'My skin is often red and inflamed. ',
+'I enjoy just about everything I do. ',
+'I never weep or feel like weeping. ',
+'I don\'t mind having someone tell me what to do. ',
+'I know of no excuse for taking advantage of someone of the opposite sex. ',
+'I feel that I am in great danger from those who wish to harm me. ',
+'Other people\'s actions rarely make me anxious. ',
+'I sometimes hear voices which no one else understands. ',
+'I am careful in almost everything I do. ',
+'I try not to get involved in conversations with others. ',
+'I think my parents would have reason to be proud of me. ',
+'I have strange fears of places and things. ',
+'My back does not bother me. ',
+'Others seem to lead happier lives than I do. ',
+'I can remember a few unpleasant things about my childhood. ',
+'Slow people make me angry. ',
+'I think that I could commit a crime and get away with it. ',
+'I rarely feel that people look for my weaknesses. ',
+'I worry when a train or bus is late. ',
+'I never confuse my own thoughts with a real person talking to me. ',
+'I\'ll try almost anything regardless of the consequences. ',
+'I have a number of close friends.. ',
+'My whole life is one big mistake. ',
+'I have periods when my mind races ahead so fast that I cannot think clearly. ',
+'Whenever I am worried about something I get cramps. ',
+'I live a very satisfying and rewarding life. ',
+'I am careful not to have any bad thoughts. ',
+'My family life has been happy and free of arguments. ',
+'Most salespeople would not cheat a customer. ',
+'When people whisper, I feel they might be talking about me. ',
+'I have the ability to concentrate without my mind wandering. ',
+'I see bright pictures in my head when I don\'t want to. ',
+'I can work for a reasonable length of time without becoming bored. ',
+'I avoid speaking with people as much as I can. ',
+'I usually do most of my daily tasks quite well. ',
+'Sometimes I feel as if I am about to lose my mind. ',
+'I seldom have a cough or sore throat. ',
+'Life is extremely dull for me. ',
+'I try to avoid jobs I dislike. ',
+'Bossy people can expect an argument from me. ',
+'No one does things for nothing. ',
+'Most people treat me openly without hiding anything. ',
+'Sometimes my own thoughts scare me so much that I think I\'m going to pass out. ',
+'My memory is as good as it ever was. ',
+'I often take risks without stopping to think of the results. ',
+'I dislike going out alone. ',
+' Poeple are better off with me,',
+'I frequently experience terrible headaches. ',
+'I often have pains in odd parts of my body. ',
+'Something interesting happens to me almost every day. ',
+'I am never cross with a loved one. ',
+'I avoid quarreling with others. ',
+'Gambling has no appeal to me. ',
+'I often have the feeling that I am not liked. ',
+'I generally feel quite comfortable when being introduced to strangers. ',
+'I can\'t always decide whether a minute or an hour has passed. ',
+'I am not the type to be bored one minute and excited about something the next. ',
+'I am happier alone than when with others. ',
+'Most people find me an interesting person to talk with. ',
+'If things don\'t improve for me, I may have to do something violent or dangerous. ',
+'I hardly ever have "splitting" headaches. ',
+'I often have trouble sleeping because I feel so sad. ',
+'On some days I am more easily annoyed than on others. ',
+'I like to run my own life without interference from anyone. ',
+'I admire a successful professional thief. ',
+'I never feel like a machine that someone else plugs in and uses. ',
+'When I am startled my heart seems to skip a beat and stop. ',
+'Things don\'t appear unusually different to me right now. ',
+'I am usually somewhat restless. ',
+'I like talking to just about anyone I meet. ',
+'I do not consider myself worthy of other people\'s kindness. ',
+'There have been periods of time when I have used alcohol to excess. ',
+'Sometimes I get so dizzy I can hardly stand up. ',
+'I always look forward to a new day. ',
+'I have never lied to anyone." ',
+'I believe in obeying those in authority. ',
+'I was not regarded as a behavior problem by my school teachers. ',
+'I am greatly concerned with what people think about me. ',
+'Emergencies seldom make me nervous. ',
+'I often see shadows and think they are people or animals. ',
+'My feelings about people do not change much from day to day. ',
+'I don\'t care whether or not the people around me are my friends. ',
+'I enjoy the respect of most people who know me. ',
+'I have often used dangerous drugs or chemicals. ',
+'I generally feel warm enough. ',
+'I don\'t think things will ever get any better for me. ',
+'Occasionally I use my friends to my own advantage. ',
+'I dislike working for a person who is too strict. ',
+'Someone is always trying to trick you. ',
+'I am sure that there is no gossiping about me. ',
+'Little things often upset me. ',
+'I usually know about what time it is. ',
+'I often leave jobs unfinished. ',
+'When I am not feeling well, I like to have someone around to comfort me. ',
+'I am not a particularly kind person. ',
+'I would enjoy watching someone suffer great pain. ',
+'I often have eye strain upon completing a day\'s work. ',
+'I believe that life is worth living. ',
+'I don\'t like thinking about personal problems. ',
+'I take great pains to be careful about other people\'s feelings. ',
+'There are many things I consider wrong and wouldn\'t do. ',
+'I often feel that others are trying to keep me out of their group. ',
+'I remain calm even in the most trying situations. ',
+'I often have the feeling that imaginary things are happening to me.. ',
+'I generally make careful plans. ',
+'I am not considered sociable. ',
+'I think I would make a very good leader. ',
+'I do not care what happens to me. ',
+'My joints give me no trouble. ',
+'I feel depressed most of the time. ',
+'As a child I sometimes felt that my parents acted unfairly. ',
+'If someone hurts me, I remember it until I can get even. ',
+'People are always trying to get away with something. ',
+'No one has a magical power to control me. ',
+'I am sometimes disturbed by things that I know can\'t hurt me. ',
+'Even when left alone, I can find my way around easily. ',
+'At times I am rather careless. ',
+'I enjoy being neighborly. ',
+'I often show poor judgment about things. ',
+'I sometimes have an urge to injure myself. ',
+'I have poor blood circulation. ',
+'I am quite content with my life as it is now. ',
+'I admit my mistakes without ever trying to hide anything. ',
+'I seldom feel like hitting anyone. ',
+'I would feel very guilty if I were caught doing something wrong. ',
+'I never feel comfortable eating food prepared by others. ',
+'I don\'t worry over what might happen to me. ',
+'At times my surroundings change so much that I think I\'m somewhere else. ',
+'I never take unnecessary chances. ',
+'I don\'t feel I need other people. ',
+'I feel capable of handling many difficult jobs. ',
+'I have been in serious trouble with the law. ',
+'I have a good deal of energy. ',
+'I dislike doing anything new. ',
+'My feelings are sometimes hurt by loved ones. ',
+'I dislike being ordered around by anyone. ',
+'I would do just about anything for money. ',
+'If I fail at something I can only blame myself. ',
+'I am usually too afraid to try anything new. ',
+'I never see things that other people cannot see. ',
+'I usually say the first thing that comes into my mind. ',
+'I truly enjoy myself at social events. ',
+'People don\'t like me because I have so many faults. ',
+'I spend a great deal of time daydreaming about things that only I know. ',
+'I lose my breath easily. ',
+'I am usually a happy person. ',
+'I\'ve never let down a friend in any way. ',
+'I do not easily lose patience with others. ',
+'Most people do what they can to help others.',
+'I\'m usually the first to be blamed if something goes wrong. ',
+'Things that upset other people usually do not bother me. ',
+'I cannot separate my daydreams from the real world.',
+'I have a well thought out reason for almost everything I undertake.',
+'I like to keep my ideas to myself.',
+'I consider myself to be a generous and pleasant person.',
+'I have been planning to do away with myself.',
+'I never feel faint.',
+'Recent events have made me feel downhearted and miserable.',
+'Sometimes I deliberately avoid a person I dislike.',
+'I get very irritated when someone disagrees with me.',
+'I sometimes have fun teasing animals.',
+'No one is trying to ruin my life.',
+'When I visit a strange place I become very upset.',
+'I can easily understand simple directions.',
+'I find it exciting to drive in a fast car.',
+'I enjoy doing things with friends whenever I am able.',
+'I am not the type of person one remembers after one meeting. ',
+'I do not care for anyone very much.',
+'I get a lot of headaches.',
+'I believe that I shall have my share of good luck.',
+'I always meet my responsibilities.',
+'No one could ever say that I am hot-tempered.',
+'Sooner or later people who break the law get caught.',
+'If certain individuals had not interfered, I would be more successful today.',
+'I do not panic more quickly than the average person.',
+'Many times I can hear mysterious voices all around me.',
+'I seldom do silly things without thinking.',
+'I make little effort to meet new people.',
+'Much of what I say is worth paying attention to.',
+'I always have difficulty sleeping.'
+];
+
+export default QUESTIONAIRES;
