+++ src-tauri/Cargo.toml
@@ -17,7 +17,7 @@ tauri-build = { version = "1.5.3", features = [] }
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.7.0", features = ["path-all"] }
+tauri = { version = "1.7.0", features = [ "dialog-message", "dialog-save", "dialog-ask", "dialog-open", "path-all"] }
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem and the built-in dev server is disabled.
